+++ packages/cosmostation/src/wallet-connect/chain-wallet.ts
@@ -1,5 +1,5 @@
 import { ChainRecord, Wallet } from '@cosmos-kit/core';
-import { ChainWCV1 } from '@cosmos-kit/wcv1';
+import { ChainWCV1 } from '../../../w';
 
 export class ChainCosmostationMobile extends ChainWCV1 {
   constructor(walletInfo: Wallet, chainInfo: ChainRecord) {