+++ packages/keplr/src/wallet-connect/client.ts
@@ -6,10 +6,10 @@ import {
 } from '@cosmjs/amino';
 import { Algo } from '@cosmjs/proto-signing';
 import { OS, SignOptions } from '@cosmos-kit/core';
-import { IWCClientV1 } from '@cosmos-kit/wcv1';
+import { IWCClientV1 } from '../../../w';
 import { KeplrIntereactionOptions } from '@keplr-wallet/types';
 import WalletConnect from '@walletconnect/client';
-import { IConnector } from '@walletconnect/types';
+import { IConnector } from '../../../w/node_modules/@walletconnect/types';
 import { payloadId, saveMobileLinkInfo } from '@walletconnect/utils';
 import deepmerge from 'deepmerge';