+++ packages/omni/src/wallet-connect/chain-wallet.ts
@@ -1,5 +1,5 @@
 import { ChainRecord, Wallet } from '@cosmos-kit/core';
-import { ChainWCV2 } from '@cosmos-kit/wcv2';
+import { ChainWCV2 } from '@cosmos-kit/walletconnect-v2';
 
 export class ChainOmniMobile extends ChainWCV2 {
   constructor(walletInfo: Wallet, chainInfo: ChainRecord) {