+++ packages/walletconnect-v2/src/registry.ts
@@ -3,7 +3,8 @@ import { Wallet } from '@cosmos-kit/core';
 export const wcv2Info: Wallet = {
   name: 'wallet-connect-v2',
   prettyName: 'Wallet Connect V2',
-  logo: 'https://docs.walletconnect.com/img/walletconnect-logo.svg',
+  logo:
+    'https://raw.githubusercontent.com/WalletConnect/walletconnect-assets/c6e9d7ca2e81d4094e83849f560a024962a7987a/Icon/Blue%20(Default)/Icon.svg',
   mode: 'wallet-connect',
   mobileDisabled: false,
 };